Added Tag and DueDate props.

+++ ApplicationCore/Services/TodoListServices.cs
@@ -19,7 +19,7 @@ namespace ApplicationCore.Services
             _mapper = mapper;
         }
 
-        public async Task<IEnumerable<TodoListDto>> GetAllListsAsync()
+        public async Task<IEnumerable<TodoListDto>> GetAllTodoListsAsync()
         {
             List<TodoListDto> resultList = new List<TodoListDto>();
             IEnumerable<Todo> todoEntities = new List<Todo>();
@@ -28,7 +28,7 @@ namespace ApplicationCore.Services
             foreach (TodoList list in await _unitOfWork.TodoListsRepository.GetAllTodoListAsync())
             {
                 TodoListDto listDto = _mapper.ToDto<TodoListDto>(list);
-                
+
                 todoEntities = await _unitOfWork.TodoListsRepository.GetAllTodoAsync(list.Id);
                 todoDtos = _mapper.ToDtoList<TodoDto>(todoEntities);
                 listDto.Todos = todoDtos;
@@ -38,33 +38,88 @@ namespace ApplicationCore.Services
             return resultList;
         }
 
-        public async Task AddAsync(TodoListDto newTodoList)
+        public async Task AddNewTodoListAsync(TodoListDto newTodoList)
         {
             TodoList list = _mapper.ToEntity<TodoList>(newTodoList);
             await _unitOfWork.TodoListsRepository.AddTodoListAsync(list);
             await _unitOfWork.SaveAsync();
         }
 
-        public async Task UpdateColorAsync(string id, string hexValue)
+        public async Task UpdateTodoListColorAsync(string id, string hexValue)
         {
-            Guid guid = Guid.Parse(id);
-            TodoList? list = await _unitOfWork.TodoListsRepository.GetTodoListByIdAsync(guid);
-            Guard.AgainstNull(list, $"Can't find todo list with id: {guid}.");
-       
-            list!.SetColor(hexValue);
+            try
+            {
+                if (!Guid.TryParse(id, out _guid))
+                {
+                    throw new ArgumentException("Invalid id format.");
+                }
+                TodoList? list = await _unitOfWork.TodoListsRepository.GetTodoListByIdAsync(_guid);
+                Guard.AgainstNull(list, $"Can't find todo list with id: {_guid}.");
 
-            _unitOfWork.TodoListsRepository.Update<TodoList>(list);
-            await _unitOfWork.SaveAsync();
+                list!.SetColor(hexValue);
+
+                _unitOfWork.TodoListsRepository.Update(list);
+                await _unitOfWork.SaveAsync();
+            }
+            catch (ArgumentNullException e)
+            {
+                throw new ArgumentNullException($"Can't find todo list with id {_guid}. {e.Message}");
+            }
+            catch (ArgumentException e)
+            {
+                throw new ArgumentException("Invalid id format.", e.Message);
+            }
         }
 
-        public async Task RemoveByIdAsync(string id)
+        public async Task UpdateTodoListTagAsync(string id, string tag)
         {
-            Guid guid = Guid.Parse(id);
-            TodoList? list = await _unitOfWork.TodoListsRepository.GetTodoListByIdAsync(guid);
-            Guard.AgainstNull(list, $"Can't find todo list with id: {guid}.");
+            try
+            {
+                if (!Guid.TryParse(id, out _guid))
+                {
+                    throw new ArgumentException("Invalid id format.");
+                }
+                TodoList? list = await _unitOfWork.TodoListsRepository.GetTodoListByIdAsync(_guid);
+                Guard.AgainstNull(list, $"Can't find todo list with id: {_guid}.");
 
-            _unitOfWork.TodoListsRepository.DeleteTodoList(list!);
-            await _unitOfWork.SaveAsync();
+                list!.Tag = tag;
+
+                _unitOfWork.TodoListsRepository.Update(list);
+                await _unitOfWork.SaveAsync();
+            }
+            catch (ArgumentNullException e)
+            {
+                throw new ArgumentNullException($"Can't find todo list with id {_guid}. {e.Message}");
+            }
+            catch (ArgumentException e)
+            {
+                throw new ArgumentException("Invalid id format.", e.Message);
+            }
         }
+
+        public async Task RemoveTodoListByIdAsync(string id)
+        {
+            try
+            {
+                if (!Guid.TryParse(id, out _guid))
+                {
+                    throw new ArgumentException("Invalid id format.");
+                }
+                TodoList? list = await _unitOfWork.TodoListsRepository.GetTodoListByIdAsync(_guid);
+                Guard.AgainstNull(list, $"Can't find todo list with id: {_guid}.");
+
+                _unitOfWork.TodoListsRepository.DeleteTodoList(list!);
+                await _unitOfWork.SaveAsync();
+            }
+            catch (ArgumentNullException e)
+            {
+                throw new ArgumentNullException($"Can't find todo list with id {_guid}. {e.Message}");
+            }
+            catch (ArgumentException e)
+            {
+                throw new ArgumentException("Invalid id format.", e.Message);
+            }
+        }
+
     }
 }
